fix(auth): one credential schema for both Auth implementations
Codex round-2 HIGH. `auth/index.ts` and `auth/service.ts` each declared their own
`Info` schema and they had silently diverged: service.ts's `Api` had no `metadata`
field. Decoding narrows a value to the declared shape, and BOTH implementations
read-modify-write the whole file, so adding or removing any unrelated provider
through `AuthService` rewrote every other entry through the narrow schema and
dropped `metadata` from all of them.

For the free tier that is not cosmetic. `metadata` carries `install_secret` — the
value the gateway hashes to derive its budget principal — and `base_url`. Losing it
disables the provider AND loses the rotation identity, so the next registration
mints a SECOND principal with its own grant rather than rotating the first.

A pre-existing divergence our feature happened to start depending on. Invisible
from either file alone, which is why two review rounds went past it; it only
surfaced once the two implementations were being compared directly.

The schema now lives in `auth/schema.ts` and both import it, so they cannot drift
again. Public surface is unchanged — index.ts re-exports the same names. service.ts
also picks up the stricter `NonNegativeInt` expiry that index.ts already applied,
which makes the two agree on what a decodable credential is rather than leaving one
stricter than the other.

Three tests, and note the third: reading back only through index.ts does NOT
discriminate, because `set` serialises the caller's object as given so metadata
reaches the file regardless — the loss happens on DECODE. It has to be read back
through service.ts too. The first version of that test passed against the bug.
